+++ web/documentserver-example/go/server/api/default/reference.go
@@ -23,11 +23,13 @@ import (
 	"net/http"
 	"net/url"
 	"strings"
+	"time"
 
 	"github.com/ONLYOFFICE/document-server-integration/server/managers"
 	"github.com/ONLYOFFICE/document-server-integration/server/models"
 	"github.com/ONLYOFFICE/document-server-integration/server/shared"
 	"github.com/ONLYOFFICE/document-server-integration/utils"
+	"github.com/golang-jwt/jwt"
 )
 
 func (srv *DefaultServerEndpointsHandler) Reference(w http.ResponseWriter, r *http.Request) {
@@ -99,6 +101,10 @@ func (srv *DefaultServerEndpointsHandler) Reference(w http.ResponseWriter, r *ht
 		},
 		Link: remoteAddr + "/editor?filename=" + url.QueryEscape(fileName),
 		Path: fileName,
+		StandardClaims: jwt.StandardClaims{
+			ExpiresAt: time.Now().Add(time.Minute * srv.config.JwtExpiresIn).Unix(),
+			IssuedAt:  time.Now().Unix(),
+		},
 	}
 
 	secret := strings.TrimSpace(srv.config.JwtSecret)